changed OtM relation direction from restaurant to location

+++ TUS-WAD-Lab4-2a-restaurantsLocations/src/main/java/week4/lab2a/tuswadlab42a/service/RestaurantService.java
@@ -3,6 +3,7 @@ package week4.lab2a.tuswadlab42a.service;
 import org.springframework.stereotype.Service;
 import week4.lab2a.tuswadlab42a.domain.Location;
 import week4.lab2a.tuswadlab42a.domain.Restaurant;
+import week4.lab2a.tuswadlab42a.repository.LocationRepository;
 import week4.lab2a.tuswadlab42a.repository.RestaurantRepository;
 
 import java.util.List;
@@ -10,8 +11,10 @@ import java.util.List;
 @Service
 public class RestaurantService {
     private final RestaurantRepository restaurantRepository;
-    public RestaurantService(RestaurantRepository restaurantRepository) {
+    private final LocationRepository locationRepository;
+    public RestaurantService(RestaurantRepository restaurantRepository, LocationRepository locationRepository) {
         this.restaurantRepository = restaurantRepository;
+        this.locationRepository = locationRepository;
     }
 
     public List<Restaurant> findAll() {
@@ -22,21 +25,23 @@ public class RestaurantService {
         return restaurantRepository.getReferenceById(id);
     }
 
-    public List<Restaurant> findByLocationId(Long locationId) {
-        return restaurantRepository.findByLocationLocationId(locationId);
-    }
-
     public Restaurant createRestaurant(Restaurant restaurant) {
-        System.out.println(restaurant.getLocation());
         return restaurantRepository.save(restaurant);
     }
 
     public Restaurant updateRestaurant(Restaurant restaurant) {
+
         return restaurantRepository.save(restaurant);
     }
 
-    public void deleteRestaurant(Long restaurantId) {
-        Restaurant restaurant = findById(restaurantId);
-        restaurantRepository.delete(restaurant);
+    public String deleteRestaurant(Long id) {
+        Restaurant restaurant = restaurantRepository.getReferenceById(id);
+        // check if no locations in restaurant
+        List<Location> locations = locationRepository.findByRestaurantId(id);
+        if (locations.isEmpty()) {
+            restaurantRepository.delete(restaurant);
+            return "Restaurant deleted";
+        }
+        return "Restaurant not deleted, locations inside";
     }
 }
